+++ src/tcerror.h
@@ -192,6 +192,7 @@ DEFINE_TCError(UnwrapType,
 
 DEFINE_TCWarn(MeaninglessCast, "meaningless cast op");
 DEFINE_TCWarn(VarShadowing, "`%s' hides already defined name of outer scope");
+DEFINE_TCWarn(TypeAliasShadowing, "`%s' hides already defined type alias of outer scope");
 
 TypeCheckError createTCErrorImpl(const Node &node, const char *kind, const char *fmt, ...)
     __attribute__((format(printf, 3, 4)));